feat(refactor): use local themes
