--- src/styles/index.scss
@@ -1,9 +1,0 @@
-// Copyright 2023 @paritytech/polkadot-staking-dashboard authors & contributors
-// SPDX-License-Identifier: GPL-3.0-only
-
-html {
-  font-size: 10.5px;
-  @media (min-width: 600px) {
-    font-size: 11px;
-  }
-}
