+++ src/theme/index.css
@@ -1,0 +1,111 @@
+/* Copyright 2023 @paritytech/polkadot-staking-dashboard authors & contributors
+SPDX-License-Identifier: GPL-3.0-only */
+
+/* NOTE: This file should be imported in the main library entry file. */
+
+html {
+  font-size: 10.5px;
+  @media (min-width: 600px) {
+    font-size: 11px;
+  }
+}
+
+/* Every element conforms to border-box by default. */
+
+* {
+  box-sizing: border-box;
+}
+
+/* Define default body style and cancel out spacing. */
+
+body {
+  font-family: Inter, sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  overflow-x: hidden;
+  font-weight: normal;
+  margin: 0;
+}
+
+/* Define default header tag styling. */
+
+h1 {
+  font-family: InterSemiBold, sans-serif;
+  font-size: 1.6rem;
+  line-height: 1.6rem;
+  font-weight: 600;
+  margin: 0;
+}
+
+h2 {
+  font-family: InterSemiBold, sans-serif;
+  font-size: 1.4rem;
+  line-height: 1.6rem;
+  font-weight: 600;
+  margin: 0;
+}
+
+h3 {
+  font-family: InterSemiBold, sans-serif;
+  font-size: 1.25rem;
+  line-height: 1.55rem;
+  font-weight: 500;
+  margin: 0;
+}
+
+h4 {
+  font-family: InterSemiBold, sans-serif;
+  font-size: 1.08rem;
+  line-height: 1.65rem;
+  font-weight: 500;
+  margin: 0;
+}
+
+h5 {
+  font-size: 0.92rem;
+  line-height: 1.02rem;
+  font-weight: 500;
+  margin: 0;
+}
+
+/* Define default paragraph style. */
+
+p {
+  font-size: 1.04rem;
+  line-height: 1.5rem;
+  margin: 0.75rem 0;
+}
+
+/* Define default anchor style. */
+
+a {
+  text-decoration: none;
+  cursor: pointer;
+}
+
+/* Define bare-bones button style. */
+
+button {
+  -webkit-tap-highlight-color: transparent;
+  font-family: Inter, sans-serif;
+  background: none;
+  border: none;
+  cursor: pointer;
+  margin: 0;
+  padding: 0;
+}
+
+/* Define default input style. */
+
+input {
+  font-family: Inter, sans-serif;
+  background: none;
+  border: none;
+}
+
+button:focus,
+input:focus,
+select:focus,
+textarea:focus {
+  outline: none;
+}
